Improved calibration method
Instead of calibrating the temperature sensor on startup and each time the settings change, now it is calibrated once for all temperatures from 10C to 30C and the values are saved to EEPROM.

At startup, the EEPROM is checked, if the data is there it is read. If it's not there the unit goes through calibration and then writes the results to EEPROM.

Normally a unit would only have to go through calibration once.